Messy: work on intro animation.
This commit is messy because I'm trying to debug a weird crash that seems to be
happening on the call p:read() in engine_util.lua. At first glance, it looks
like a bug in the implementation of read on the result of an io.popen() call. I
can investigate further by pulling in the lua 5.3.3 source and trying to
pinpoint the error with Lua's source — if, indeed, there is even an error in
Lua's source.